+++ inst/app/tools/analysis/crtree_ui.R
@@ -152,32 +152,16 @@ output$ui_crtree_store_res_name <- renderUI({
   textInput("crtree_store_res_name", "Store residuals:", "", placeholder = "Provide variable name")
 })
 
-observe({
-  ## dep on most inputs
-  input$data_filter
-  input$show_filter
-  sapply(r_drop(names(crtree_args)), function(x) input[[paste0("crtree_", x)]])
-
-  ## notify user when the model needs to be updated
-  ## based on https://stackoverflow.com/questions/45478521/listen-to-reactive-invalidation-in-shiny
-  if (pressed(input$crtree_run)) {
-    if (is.null(input$crtree_evar)) {
-      updateTabsetPanel(session, "tabs_crtree ", selected = "Summary")
-      updateActionButton(session, "crtee_run", "Estimate model", icon = icon("play"))
-    } else
-    if (isTRUE(attr(crtree_inputs, "observable")$.invalidated)) {
-      updateActionButton(session, "crtree_run", "Re-estimate model", icon = icon("refresh", class = "fa-spin"))
-    } else {
-      updateActionButton(session, "crtree_run", "Estimate model", icon = icon("play"))
-    }
-  }
-})
+## add a spinning refresh icon if the model needs to be (re)estimated
+run_refresh(crtree_args, "crtree", tabs = "tabs_crtree", label = "Estimate model", relabel = "Re-estimate model")
 
 output$ui_crtree <- renderUI({
   req(input$dataset)
   tagList(
-    wellPanel(
-      actionButton("crtree_run", "Estimate model", width = "100%", icon = icon("play"), class = "btn-success")
+    conditionalPanel(condition = "input.tabs_crtree == 'Summary'",
+      wellPanel(
+        actionButton("crtree_run", "Estimate model", width = "100%", icon = icon("play"), class = "btn-success")
+      )
     ),
     wellPanel(
       conditionalPanel(