+++ inst/app/tools/analysis/logistic_ui.R
@@ -274,31 +274,16 @@ output$ui_logit_store_res_name <- renderUI({
   textInput("logit_store_res_name", "Store residuals:", "", placeholder = "Provide variable name")
 })
 
-observe({
-  ## dep on most inputs
-  input$data_filter
-  input$show_filter
-  sapply(r_drop(names(logit_args)), function(x) input[[paste0("logit_", x)]])
-
-  ## notify user when the model needs to be updated
-  ## based on https://stackoverflow.com/questions/45478521/listen-to-reactive-invalidation-in-shiny
-  if (pressed(input$logit_run)) {
-    if (is.null(input$logit_evar)) {
-      updateTabsetPanel(session, "tabs_logistic ", selected = "Summary")
-      updateActionButton(session, "logit_run", "Estimate model", icon = icon("play"))
-    } else if (isTRUE(attr(logit_inputs, "observable")$.invalidated)) {
-      updateActionButton(session, "logit_run", "Re-estimate model", icon = icon("refresh", class = "fa-spin"))
-    } else {
-      updateActionButton(session, "logit_run", "Estimate model", icon = icon("play"))
-    }
-  }
-})
+## add a spinning refresh icon if the model needs to be (re)estimated
+run_refresh(logit_args, "logit", tabs = "tabs_logistic", label = "Estimate model", relabel = "Re-estimate model")
 
 output$ui_logistic <- renderUI({
   req(input$dataset)
   tagList(
-    wellPanel(
-      actionButton("logit_run", "Estimate model", width = "100%", icon = icon("play"), class = "btn-success")
+    conditionalPanel(condition = "input.tabs_logistic == 'Summary'",
+      wellPanel(
+        actionButton("logit_run", "Estimate model", width = "100%", icon = icon("play"), class = "btn-success")
+      )
     ),
     wellPanel(
       conditionalPanel(