+++ inst/app/tools/analysis/mnl_ui.R
@@ -292,31 +292,17 @@ output$ui_mnl_store_res_name <- renderUI({
   textInput("mnl_store_res_name", "Store residuals:", "", placeholder = "Provide variable name")
 })
 
-observe({
-  ## dep on most inputs
-  input$data_filter
-  input$show_filter
-  sapply(r_drop(names(mnl_args)), function(x) input[[paste0("mnl_", x)]])
-
-  ## notify user when the model needs to be updated
-  ## based on https://stackoverflow.com/questions/45478521/listen-to-reactive-invalidation-in-shiny
-  if (pressed(input$mnl_run)) {
-    if (is.null(input$mnl_evar)) {
-      updateTabsetPanel(session, "tabs_mnl ", selected = "Summary")
-      updateActionButton(session, "mnl_run", "Estimate model", icon = icon("play"))
-    } else if (isTRUE(attr(mnl_inputs, "observable")$.invalidated)) {
-      updateActionButton(session, "mnl_run", "Re-estimate model", icon = icon("refresh", class = "fa-spin"))
-    } else {
-      updateActionButton(session, "mnl_run", "Estimate model", icon = icon("play"))
-    }
-  }
-})
+## add a spinning refresh icon if the model needs to be (re)estimated
+run_refresh(reg_args, "mnl", tabs = "tabs_mnl", label = "Estimate model", relabel = "Re-estimate model")
 
 output$ui_mnl <- renderUI({
   req(input$dataset)
   tagList(
-    wellPanel(
-      actionButton("mnl_run", "Estimate model", width = "100%", icon = icon("play"), class = "btn-success")
+    conditionalPanel(
+      condition = "input.tabs_mnl == 'Summary'",
+      wellPanel(
+        actionButton("mnl_run", "Estimate model", width = "100%", icon = icon("play"), class = "btn-success")
+      )
     ),
     wellPanel(
       conditionalPanel(